+++ exporters/otlp/otlptrace/otlptracegrpc/options.go
@@ -60,7 +60,7 @@ func WithInsecure() Option {
 // If the OTEL_EXPORTER_OTLP_ENDPOINT or OTEL_EXPORTER_OTLP_TRACES_ENDPOINT
 // environment variable is set, and this option is not passed, that variable
 // value will be used. If both environment variables are set,
-// OTEL_EXPORTER_OTLP_TRACES_ENDPOINT will take precedence. If the environment
+// OTEL_EXPORTER_OTLP_TRACES_ENDPOINT will take precedence. If an environment
 // variable is set, and this option is passed, this option will take precedence.
 //
 // If both this option and WithEndpointURL are used, the last used option will
@@ -81,7 +81,7 @@ func WithEndpoint(endpoint string) Option {
 // If the OTEL_EXPORTER_OTLP_ENDPOINT or OTEL_EXPORTER_OTLP_TRACES_ENDPOINT
 // environment variable is set, and this option is not passed, that variable
 // value will be used. If both environment variables are set,
-// OTEL_EXPORTER_OTLP_TRACES_ENDPOINT will take precedence. If the environment
+// OTEL_EXPORTER_OTLP_TRACES_ENDPOINT will take precedence. If an environment
 // variable is set, and this option is passed, this option will take precedence.
 //
 // If both this option and WithEndpoint are used, the last used option will